2024年11月18日 16:34

+++ Test_Learning/hw_lab/第四次实验/lab13.c
@@ -3,7 +3,7 @@
 typedef struct {
     char name[100];
     char stu_num[100];
-    int scores;
+    double scores;
 } Stu;
 
 void sort (Stu* ArrStu);
